backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
langchain
langchain-groq
streamlit